made new interf and moved code around
